+++ test-jest/methods.test.js
@@ -4,28 +4,42 @@ import envFactory from '../__tests__/Store/envFactory';
 import Hints from '../__tests__/Modules/Hints/Resource/resource';
 
 const mock = new MockAdapter(axios);
-mock.onGet().reply(403);
 
-describe('Methods: Update', () => {
-  it('Should return the network error when it fails', async () => {
-    const {registerResource, store} = envFactory({
+describe('Axios Error', () => {
+  it('Should return 403 status error when it fails due to authentication', async () => {
+    mock.onGet().reply(403);
+
+    const {registerResource} = envFactory({
       errorHandler: (error) => {
-        Promise.reject(error);
+        throw new Error(JSON.stringify(error));
       },
     });
     const seenHintsResource = registerResource(Hints).Hints;
-    const initialHints = store.getters['Hints/hints'];
-
-    expect(initialHints.length).toBe(0);
-
-    const id = 1;
 
     try {
-      await seenHintsResource.get(null, id);
-    } catch (error) {
-      console.info(error);
-      console.info(error.internalError);
-      expect(error.internalError.response.status).toBe('403');
+      await seenHintsResource.get(null, 1);
+    } catch (err) {
+      const error = JSON.parse(err.message);
+      expect(error.internalError.response.status).toBe(403);
+    }
+  });
+
+  it('Should return network status error when it fails without mocking', async () => {
+    mock.restore();
+
+    const {registerResource} = envFactory({
+      errorHandler: (error) => {
+        throw new Error(JSON.stringify(error));
+      },
+    });
+    const seenHintsResource = registerResource(Hints).Hints;
+
+    try {
+      await seenHintsResource.get(null, 1);
+    } catch (err) {
+      const error = JSON.parse(err.message);
+      expect(error.internalError.response).toBeUndefined;
+      expect(error.internalError.message).toBe('Network Error');
     }
   });
 });
